Introduce a compatible class to handle walk and find method
